refactor: bump leansig version and ssz dependencies
