+++ Cargo.toml
@@ -92,16 +92,16 @@ discv5 = { version = "0.10.2", features = ["libp2p"] }
 enr = "0.13.0"
 ethereum_hashing = { git = "https://github.com/ReamLabs/ethereum_hashing.git", rev = "c14a9ad96436e6b8b631faf49a3666e4558b71ae" }
 ethereum_serde_utils = "0.8"
-ethereum_ssz = "0.9.1"
-ethereum_ssz_derive = "0.9.1"
+ethereum_ssz = "0.10"
+ethereum_ssz_derive = "0.10"
 eventsource-client = "0.15.1"
 futures = "0.3"
 hashbrown = "0.16.0"
-hashsig = { git = "https://github.com/leanEthereum/leanSig.git", rev = "2bd2b422dd225aa211fa37a1589b1fa55acb7992" }
 itertools = "0.14"
 jsonwebtoken = "9.3.1"
 kzg = { git = "https://github.com/grandinetech/rust-kzg" }
 lazy_static = "1.5.0"
+leansig = { git = "https://github.com/leanEthereum/leanSig.git", rev = "0be24355e2872464fb38b7a6f91bfe63e0055d9e" }
 libp2p = { version = "0.56", default-features = false, features = ["identify", "yamux", "noise", "dns", "serde", "tcp", "tokio", "plaintext", "secp256k1", "macros", "ecdsa", "metrics", "quic", "upnp", "gossipsub", "ping"] }
 libp2p-identity = "0.2.12"
 libp2p-mplex = "0.43.1"
@@ -119,7 +119,7 @@ serde_json = "1.0.145"
 serde_yaml = "0.9"
 sha2 = "0.10.9"
 snap = "1.1"
-ssz_types = { git = "https://github.com/ReamLabs/ssz_types", branch = "removable-variable-list" }
+ssz_types = { git = "https://github.com/sigp/ssz_types.git", rev = "5dc2d521ed31f8e38b0065a80bf98af55aac06f8" }
 tempdir = "0.3.7"
 tempfile = "3.23.0"
 thiserror = "2.0.17"
@@ -128,8 +128,8 @@ tokio-util = { version = "0.7.16", features = ["compat"] }
 tracing = "0.1"
 tracing-subscriber = "0.3.20"
 tracing-test = "0.2.5"
-tree_hash = "0.10"
-tree_hash_derive = "0.11.0"
+tree_hash = "0.12"
+tree_hash_derive = "0.12"
 unicode-normalization = "0.1.24"
 url = "2.5.7"
