+++ bin/ream/Cargo.toml
@@ -21,7 +21,7 @@ bip39.workspace = true
 clap = { workspace = true, features = ["derive", "env"] }
 discv5.workspace = true
 hashbrown.workspace = true
-hashsig.workspace = true
+leansig.workspace = true
 libp2p-identity.workspace = true
 prometheus_exporter.workspace = true
 rand.workspace = true
