+++ crates/common/consensus/lean/src/block.rs
@@ -1,6 +1,6 @@
 use alloy_primitives::B256;
 use anyhow::{anyhow, ensure};
-use ream_post_quantum_crypto::hashsig::signature::Signature;
+use ream_post_quantum_crypto::leansig::signature::Signature;
 use serde::{Deserialize, Serialize};
 use ssz_derive::{Decode, Encode};
 use ssz_types::{VariableList, typenum::U4096};
