+++ crates/crypto/bls/src/signature.rs
@@ -27,7 +27,7 @@ impl<'de> Deserialize<'de> for BLSSignature {
     {
         let result: String = Deserialize::deserialize(deserializer)?;
         let result = hex::decode(&result).map_err(serde::de::Error::custom)?;
-        let key = FixedVector::from(result);
+        let key = FixedVector::try_from(result).map_err(serde::de::Error::custom)?;
         Ok(Self { inner: key })
     }
 }
@@ -39,7 +39,7 @@ impl BLSSignature {
 
     pub fn infinity() -> Self {
         Self {
-            inner: FixedVector::from(vec![
+            inner: FixedVector::try_from(vec![
                 0xc0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
                 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
                 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
@@ -47,7 +47,8 @@ impl BLSSignature {
                 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
                 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
                 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            ]),
+            ])
+            .expect("This is supposed to be right"),
         }
     }
 }
