+++ crates/crypto/keystore/Cargo.toml
@@ -13,7 +13,7 @@ version.workspace = true
 aes.workspace = true
 alloy-primitives.workspace = true
 anyhow.workspace = true
-hashsig.workspace = true
+leansig.workspace = true
 rand.workspace = true
 serde.workspace = true
 serde_json.workspace = true
