+++ crates/crypto/post_quantum/Cargo.toml
@@ -21,7 +21,7 @@ anyhow.workspace = true
 bincode.workspace = true
 ethereum_ssz.workspace = true
 ethereum_ssz_derive.workspace = true
-hashsig.workspace = true
+leansig.workspace = true
 rand.workspace = true
 serde.workspace = true
 serde_json.workspace = true
