+++ crates/crypto/post_quantum/src/leansig/mod.rs
@@ -6,13 +6,13 @@ pub mod signature;
 use bincode::config::{Fixint, LittleEndian, NoLimit};
 
 #[cfg(all(not(test), feature = "signature-scheme-prod"))]
-pub type HashSigScheme = hashsig::signature::generalized_xmss::instantiations_poseidon_top_level::lifetime_2_to_the_32::hashing_optimized::SIGTopLevelTargetSumLifetime32Dim64Base8;
+pub type HashSigScheme = leansig::signature::generalized_xmss::instantiations_poseidon_top_level::lifetime_2_to_the_32::hashing_optimized::SIGTopLevelTargetSumLifetime32Dim64Base8;
 
 #[cfg(all(not(test), feature = "signature-scheme-test"))]
-pub type HashSigScheme = hashsig::signature::generalized_xmss::instantiations_poseidon_top_level::lifetime_2_to_the_8::SIGTopLevelTargetSumLifetime8Dim64Base8;
+pub type HashSigScheme = leansig::signature::generalized_xmss::instantiations_poseidon_top_level::lifetime_2_to_the_8::SIGTopLevelTargetSumLifetime8Dim64Base8;
 
 #[cfg(test)]
-pub type HashSigScheme = hashsig::signature::generalized_xmss::instantiations_poseidon_top_level::lifetime_2_to_the_8::SIGTopLevelTargetSumLifetime8Dim64Base8;
+pub type HashSigScheme = leansig::signature::generalized_xmss::instantiations_poseidon_top_level::lifetime_2_to_the_8::SIGTopLevelTargetSumLifetime8Dim64Base8;
 
 /// NOTE: `GeneralizedXMSSSignature` doesn't implement methods like `to_bytes`,
 /// which means we need to use bincode to serialize/deserialize it.
