+++ crates/crypto/post_quantum/src/lib.rs
@@ -1,11 +1,1 @@
-pub mod hashsig;
-
-use serde::{Deserialize, Serialize};
-use ssz_derive::{Decode, Encode};
-use tree_hash_derive::TreeHash;
-
-/// Stub PQ signature struct to avoid issues with Consensus crate
-#[derive(
-    Debug, PartialEq, Eq, Clone, Serialize, Deserialize, Encode, Decode, TreeHash, Default,
-)]
-pub struct PQSignature {}
+pub mod leansig;
